+++ model/src/main/java/model/dao/BusRouteMaximalDAO.java
@@ -5,6 +5,6 @@ import javax.ejb.Local;
 import model.BusRouteMaximal;
 
 @Local
-public interface BusRouteMaximalDAO extends ModelDAO<BusRouteMaximal, Integer> {
+public interface BusRouteMaximalDAO extends ModelDAO<BusRouteMaximal, Long> {
 
 }
