+++ model/src/main/java/model/dao/CornerDAO.java
@@ -9,7 +9,7 @@ import model.Corner;
 import com.vividsolutions.jts.geom.Point;
 
 @Local
-public interface CornerDAO extends ModelDAO<Corner, Integer> {
+public interface CornerDAO extends ModelDAO<Corner, Long> {
 
 	public List<Corner> findCornersWithinRadius(Point point,Double radius);
